test(tools): share the admitted-tool harness across effect-gate suites (#2190)
The refusal-session-gate suite stood up its own admitting session: the same 46
lines the read-only recovery suite already carried, differing only in the run
and session ids. SonarCloud read the new code as 21.9% duplicated against a 3%
gate.

Both copies were fixture setup, not assertions: an admission client that allows
every call, a session that collects the events the effect gate files, and a
permission-mode registry. `bindAdmittedToolHarness` is that setup once, with the
admission scope named by a `label` so each suite keeps its own run/session/turn
ids and the read-only suite still passes its own permission context through.

No assertion moved, no case merged, no scenario dropped. Reverting the source
still fails the same three tests (the sweep plus both refusal cases), and
marking web_fetch side-effecting still fails the same six read-only cases, so
the harness detects gate poisoning exactly as the inline copies did.
